refacror: move data structure files
