+++ lib/ui/timer/model/timer_view_state.freezed.dart
